start-qemu.sh: remove "tdx_disable_filter" requirement
When using vsock to get TDREPORT, passing "tdx_disable_filter" on the kernel
command line used to be required as a workaround. However, this workaround also
compromised the security.

Disabling the filter is no longer needed for recent kernels.
Vsock driver has been audited and added to the list of allowed TDX devices.

Signed-off-by: Juro Bystricky <juro.bystricky@intel.com>